Improve efficiency of gather existing property step in property writer
Multiple queries were being made to the database, when the intention was that all data was gathered
in a single JOINed query. This has been resolved, at the cost of currently not being able to include
links in the request for data. Such links were implemented in a ill-defined way, and probably unused
for practical purposes.