Address issues #123 and #124: earliest() and latest() live calculations
* latest() would give no result if called from the last snapshot.
* earliest() would give no result if the halo did not exist in the very first timestep.

The problems were quite deeply embedded in the algorithm, which looked for the corresponding halo in a particular
timestep instead of just following the tree as far as possible (as other parts of tangos would do).

To address this, both functions have been moved to a new algorithm based on the existing find_progenitor function.