Permit use of preloop in live calculations
Preloop is a mechanism by which a property can pre-cache information about the timestep
on which it is running. This is essential for many properties, but was quietly
un-implemented for live-calculation. For example, if you want to know the relationship
of a halo to all others in a given timestep, a preloop step where information about
the population is cached is essential for efficiency.