Attempt to fix AHF long integer issue
Some versions of AHF generates internal unsigned long halo numbers,
but SQLite can only store signed longs (8-bit maximum). This leads
to overflow errors when generating a database.

Since ordering is unimportant, a hack to fix this may be to take the
bit pattern and turn it into a signed integer before storing.