Be more aggressive about removing links to the live database
While not caught by tests, on real-world scenarios we are seeing deep sqlalchemy
errors when trying to write properties using the new architecture, seemingly
because the pickled objects are carrying more state with them than they should.

This is an attempt to be more aggressive about getting everything into a transient
state. It breaks some tests, so is experimental.